refactor(mobile): partir inicio_tab.dart en 3 modulos (fase 5b-3)
- inicio_tab.dart        202  pantalla + _SubTituloFecha + _SeccionTitulo
- card_proxima_cita.dart 215  CardProximaCita + CardSinCita + dialog cancelar
- inicio_widgets.dart    111  MiniStat + QuickAccess

Archivo original (516 lineas) movido a subcarpeta inicio/.
Sin cambios funcionales. flutter analyze: 0 issues.
